refactoring the project structure
