+++ bench-uplc/Scripts.hs
@@ -1,9 +1,9 @@
-{-# LANGUAGE TemplateHaskell   #-}
+{-# LANGUAGE TemplateHaskell #-}
 
 {-# OPTIONS_GHC -fplugin-opt PlutusTx.Plugin:profile-all #-}
 {-# OPTIONS_GHC -fplugin-opt PlutusTx.Plugin:conservative-optimisation #-}
 
-module Script (compiledSymbolicVerifier, compiledPlonkVerifier, compiledPlonkVerify) where
+module Scripts (compiledSymbolicVerifier, compiledPlonkVerifier, compiledPlonkVerify) where
 
 import           PlutusLedgerApi.V3                       (ScriptContext (..))
 import           PlutusTx                                 (CompiledCode)
@@ -11,7 +11,7 @@ import           PlutusTx.Prelude                         (Bool)
 import           PlutusTx.TH                              (compile)
 
 import           ZkFold.Base.Protocol.NonInteractiveProof (NonInteractiveProof (..))
-import           ZkFold.Cardano.OnChain                   (plonkVerifier, symbolicVerifier)
+import           ZkFold.Cardano.ScriptsVerifier                   (plonkVerifier, symbolicVerifier)
 import           ZkFold.Cardano.Plonk                     (PlonkPlutus)
 
 compiledSymbolicVerifier :: CompiledCode (Setup PlonkPlutus -> Input PlonkPlutus -> Proof PlonkPlutus -> ScriptContext -> Bool)
