+++ src/ZkFold/Cardano/Plonk.hs
@@ -1,72 +1,21 @@
-{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE TypeFamilies   #-}
+{-# OPTIONS_GHC -Wno-incomplete-uni-patterns #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
 module ZkFold.Cardano.Plonk where
 
-import qualified Data.Vector                              as Vector
 import           GHC.ByteOrder                            (ByteOrder (..))
-import           GHC.Natural                              (naturalToInteger)
-import           PlutusTx.Builtins                        (BuiltinByteString, blake2b_256, bls12_381_G1_uncompress, bls12_381_G2_uncompress, bls12_381_finalVerify, bls12_381_millerLoop, byteStringToInteger, consByteString, emptyByteString, integerToByteString)
-import           PlutusTx.List                            (and, foldr, head, map, zipWith)
-import           PlutusTx.Prelude                         (Bool (..), divide, enumFromTo, even, otherwise, sum, takeByteString, ($), (&&), (.), (<), (<>), (>), (||))
-import qualified PlutusTx.Prelude                         as Plutus
-import           Prelude                                  (Integral (..), Show (..), fromInteger, undefined, (++))
+import           PlutusTx.Builtins
+import           PlutusTx.List                            (zipWith)
+import           PlutusTx.Prelude                         (Bool (..), takeByteString, ($), (.), (<>))
+import           Prelude                                  (undefined)
 
 import qualified ZkFold.Base.Algebra.Basic.Class          as ZkFold
 import           ZkFold.Base.Algebra.Basic.Class
-import           ZkFold.Base.Algebra.Basic.Number         (value)
-import           ZkFold.Base.Protocol.ARK.Plonk           hiding (F)
-import           ZkFold.Base.Protocol.NonInteractiveProof (FromTranscript (..), NonInteractiveProof (..), ToTranscript (..))
-import           ZkFold.Cardano.Plonk.Inputs              (InputBytes (..), ProofBytes (..), SetupBytes (..))
-import           ZkFold.Cardano.Plonk.Internal            (F (..), challenge, convertF, convertG1, convertG2, getXi, mul, powMod)
-
-type Plonk32 = Plonk 32 BuiltinByteString
+import           ZkFold.Base.Protocol.NonInteractiveProof (NonInteractiveProof (..))
+import           ZkFold.Cardano.Plonk.OnChain
 
 data PlonkPlutus
-
-mkSetup :: Setup Plonk32 -> Setup PlonkPlutus
-mkSetup (PlonkSetupParams {..}, _, _, PlonkCircuitCommitments {..}, PlonkInput input, _) = SetupBytes
-  { n     = naturalToInteger $ value @32
-  , g0'   = Plutus.head . Vector.toList . Vector.map convertG1 $ gs
-  , h0'   = convertG2 h0
-  , h1'   = convertG2 h1
-  , omega = F $ convertF omega
-  , k1    = F $ convertF k1
-  , k2    = F $ convertF k2
-  , cmQl' = convertG1 cmQl
-  , cmQr' = convertG1 cmQr
-  , cmQo' = convertG1 cmQo
-  , cmQm' = convertG1 cmQm
-  , cmQc' = convertG1 cmQc
-  , cmS1' = convertG1 cmS1
-  , cmS2' = convertG1 cmS2
-  , cmS3' = convertG1 cmS3
-  , gens  = map (powMod (F $ convertF omega)) (enumFromTo 1 $ toInteger $ Vector.length input)
-  }
-
-mkInput :: Input Plonk32 -> Input PlonkPlutus
-mkInput (PlonkInput input) = InputBytes . Vector.toList . Vector.map (F . convertF) $ input
-
-mkProof ::  Setup PlonkPlutus -> Proof Plonk32 -> Proof PlonkPlutus
-mkProof SetupBytes{n, gens} (PlonkProof cmA cmB cmC cmZ cmT1 cmT2 cmT3 proof1 proof2 a_xi b_xi c_xi s1_xi s2_xi z_xi) = ProofBytes
-  { cmA'    = convertG1 cmA
-  , cmB'    = convertG1 cmB
-  , cmC'    = convertG1 cmC
-  , cmZ'    = convertG1 cmZ
-  , cmT1'   = convertG1 cmT1
-  , cmT2'   = convertG1 cmT2
-  , cmT3'   = convertG1 cmT3
-  , proof1' = convertG1 proof1
-  , proof2' = convertG1 proof2
-  , a_xi'   = convertF a_xi
-  , b_xi'   = convertF b_xi
-  , c_xi'   = convertF c_xi
-  , s1_xi'  = convertF s1_xi
-  , s2_xi'  = convertF s2_xi
-  , z_xi'   = convertF z_xi
-  , lagsInv = let xi = getXi (convertG1 cmA) (convertG1 cmB) (convertG1 cmC) (convertG1 cmZ) (convertG1 cmT1) (convertG1 cmT2) (convertG1 cmT3)
-              in  map ((/) one . (\x -> F n * (xi - x))) gens
-  }
 
 instance NonInteractiveProof PlonkPlutus where
     type Transcript PlonkPlutus   = BuiltinByteString
