+++ src/ZkFold/Cardano/Plonk/OnChain.hs
@@ -1,24 +1,19 @@
 {-# LANGUAGE TemplateHaskell #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
-module ZkFold.Cardano.Plonk.Internal where
+module ZkFold.Cardano.Plonk.OnChain where
 
 import           Data.Aeson                               (FromJSON, ToJSON)
 import           GHC.ByteOrder                            (ByteOrder (..))
 import           GHC.Generics                             (Generic)
-import           GHC.Natural                              (naturalToInteger)
 import           PlutusTx                                 (makeIsDataIndexed, makeLift)
 import           PlutusTx.Builtins
 import           PlutusTx.Prelude                         hiding (fromInteger)
-import           Prelude                                  (Num (fromInteger))
+import           Prelude                                  (Num (fromInteger), Show)
 import qualified Prelude                                  as Haskell
 
 import qualified ZkFold.Base.Algebra.Basic.Class          as ZkFold
-import           ZkFold.Base.Algebra.Basic.Field          (Ext2 (..), Zp, fromZp, toZp)
-import           ZkFold.Base.Algebra.EllipticCurve.Class  (Point (..))
-import qualified ZkFold.Base.Protocol.ARK.Plonk           as Plonk
 import           ZkFold.Base.Protocol.NonInteractiveProof (FromTranscript (..), ToTranscript (..))
-
--- TODO: separate on-chain and off-chain code
 
 ---------------------------------- F --------------------------------------
 
@@ -26,7 +21,8 @@ bls12_381_field_prime :: Integer
 bls12_381_field_prime = 52435875175126190479447740508185965837690552500527637822603658699938581184513
 
 newtype F = F { toF :: Integer }
-    deriving (Haskell.Show, Generic, ToJSON, FromJSON)
+  deriving stock (Show, Generic)
+  deriving newtype (ToJSON, FromJSON)
 makeLift ''F
 makeIsDataIndexed ''F [('F,0)]
 
@@ -126,40 +122,6 @@ instance ZkFold.AdditiveGroup BuiltinBLS12_381_G2_Element where
     {-# INLINABLE (-) #-}
     g - h = bls12_381_G2_add g (bls12_381_G2_neg h)
 
--------------------------- Conversions ------------------------------------
-
-convertF :: Plonk.F -> Integer
-convertF = naturalToInteger . fromZp
-
-convertPlonkF :: F -> Plonk.F
-convertPlonkF = toZp . toF
-
-convertZp :: Zp p -> Integer
-convertZp = naturalToInteger . fromZp
-
--- See CIP-0381 for the conversion specification
-convertG1 :: Plonk.G1 -> BuiltinByteString
-convertG1 Inf = bls12_381_G1_compressed_zero
-convertG1 (Point x y) = bs
-    where
-        bsX = integerToByteString BigEndian 48 $ convertZp x
-        b   = indexByteString bsX 0
-        b'  = b + 128 + 32 * (if y Haskell.> ZkFold.negate y then 1 else 0)
-        bs  = consByteString b' $ sliceByteString 1 47 bsX
-
--- See CIP-0381 for the conversion specification
-convertG2 :: Plonk.G2 -> BuiltinByteString
-convertG2 Inf = bls12_381_G2_compressed_zero
-convertG2 (Point x y) = bs
-    where
-        f (Ext2 a0 a1) = integerToByteString BigEndian 48 (convertZp a1) <> integerToByteString BigEndian 48 (convertZp a0)
-        bsX  = f x
-        bsY  = f y
-        bsY' = f $ ZkFold.negate y
-        b   = indexByteString bsX 0
-        b'  = b + 128 + 32 * (if bsY `greaterThanByteString` bsY' then 1 else 0)
-        bs  = consByteString b' $ sliceByteString 1 95 bsX
-
 -------------------------- Transcript -------------------------------------
 
 type Transcript = BuiltinByteString
@@ -168,15 +130,9 @@ instance ToTranscript BuiltinByteString F where
     {-# INLINABLE toTranscript #-}
     toTranscript f = integerToByteString BigEndian 32 $ toF f
 
-instance ToTranscript BuiltinByteString Plonk.F where
-    toTranscript = toTranscript . F . convertZp
-
 instance ToTranscript BuiltinByteString G1 where
     {-# INLINABLE toTranscript #-}
     toTranscript = bls12_381_G1_compress
-
-instance ToTranscript BuiltinByteString Plonk.G1 where
-    toTranscript = toTranscript . bls12_381_G1_uncompress . convertG1
 
 instance FromTranscript BuiltinByteString F where
     {-# INLINABLE newTranscript #-}
@@ -185,25 +141,61 @@ instance FromTranscript BuiltinByteString F where
     {-# INLINABLE fromTranscript #-}
     fromTranscript = F . byteStringToInteger BigEndian . takeByteString 31 . blake2b_256
 
-instance FromTranscript BuiltinByteString Plonk.F where
-    newTranscript = newTranscript @BuiltinByteString @F
-
-    fromTranscript = toZp . toF . fromTranscript @BuiltinByteString @F
-
 {-# INLINABLE challenge #-}
 challenge :: Transcript -> (F, Transcript)
 challenge ts =
     let ts' = newTranscript @BuiltinByteString @F ts
     in (fromTranscript ts', ts')
 
-getXi :: BuiltinByteString -> BuiltinByteString -> BuiltinByteString ->
-         BuiltinByteString -> BuiltinByteString -> BuiltinByteString -> BuiltinByteString -> F
-getXi cmA' cmB' cmC' cmZ' cmT1' cmT2' cmT3' = xi
-  where
-    (beta, ts) = challenge $ cmA' <> cmB' <> cmC'
+---------------------------------- ByteString ----------------------------------
 
-    (gamma, t1) = challenge ts
+data SetupBytes = SetupBytes {
+    n     :: Integer
+  , g0'   :: BuiltinByteString
+  , h0'   :: BuiltinByteString
+  , h1'   :: BuiltinByteString
+  , omega :: F
+  , k1    :: F
+  , k2    :: F
+  , cmQl' :: BuiltinByteString
+  , cmQr' :: BuiltinByteString
+  , cmQo' :: BuiltinByteString
+  , cmQm' :: BuiltinByteString
+  , cmQc' :: BuiltinByteString
+  , cmS1' :: BuiltinByteString
+  , cmS2' :: BuiltinByteString
+  , cmS3' :: BuiltinByteString
+  , gens  :: [F]
+} deriving stock (Show)
 
-    (alpha, t2) = challenge $ t1 <> cmZ'
+makeLift ''SetupBytes
+makeIsDataIndexed ''SetupBytes [('SetupBytes,0)]
 
-    (xi, t3) = challenge $ t2 <> cmT1' <> cmT2' <> cmT3'
+newtype InputBytes = InputBytes {
+  pubInput :: [F]
+} deriving stock (Show)
+
+makeLift ''InputBytes
+makeIsDataIndexed ''InputBytes [('InputBytes,0)]
+
+data ProofBytes = ProofBytes {
+    cmA'    :: BuiltinByteString
+  , cmB'    :: BuiltinByteString
+  , cmC'    :: BuiltinByteString
+  , cmZ'    :: BuiltinByteString
+  , cmT1'   :: BuiltinByteString
+  , cmT2'   :: BuiltinByteString
+  , cmT3'   :: BuiltinByteString
+  , proof1' :: BuiltinByteString
+  , proof2' :: BuiltinByteString
+  , a_xi'   :: Integer
+  , b_xi'   :: Integer
+  , c_xi'   :: Integer
+  , s1_xi'  :: Integer
+  , s2_xi'  :: Integer
+  , z_xi'   :: Integer
+  , lagsInv :: [F]
+} deriving stock (Show)
+
+makeLift ''ProofBytes
+makeIsDataIndexed ''ProofBytes [('ProofBytes,0)]
