--- zkfold-cardano-benchmark/exe/Main.hs
@@ -1,9 +1,0 @@
-module Main where
-
-import           Prelude
-import           System.IO                            (stdout)
-
-import           ZkFoldBenchmark.Verifier.RunVerifier (runVerifier)
-
-main :: IO ()
-main = do runVerifier stdout
