--- zkfold-cardano-benchmark/zkfold-cardano-benchmark.cabal
@@ -1,110 +1,0 @@
-cabal-version:      3.0
-
-name:               zkfold-cardano-benchmark
-version:            0.1.0.0
-synopsis:           zero-knowledge proof protocols benchmark
-category:           Testing, Cryptography, Math, Symbolic Computation
-author:             Vladimir Sinyakov, Ilya Baryshnikov
-maintainer:         info@zkfold.io
-license:            MIT
-build-type:         Simple
-
-common lang
-  default-language:   Haskell2010
-  default-extensions:
-    DeriveFoldable
-    DeriveFunctor
-    DeriveGeneric
-    DeriveLift
-    DeriveTraversable
-    DerivingStrategies
-    DerivingVia
-    ImportQualifiedPost
-    LambdaCase
-    NumericUnderscores
-    GeneralizedNewtypeDeriving
-    StandaloneDeriving
-    ViewPatterns
-    AllowAmbiguousTypes
-    FlexibleContexts
-    RankNTypes
-    TypeApplications
-    DataKinds
-    InstanceSigs
-    MultiParamTypeClasses
-    RecordWildCards
-    ScopedTypeVariables
-    FlexibleInstances
-    NoImplicitPrelude
-    Strict
-
-  ghc-options:
-    -fno-specialise
-    -fno-spec-constr
-    -fno-ignore-interface-pragmas
-    -fno-omit-interface-pragmas
-    -fno-strictness
-    -fno-unbox-strict-fields 
-    -fno-unbox-small-strict-fields
-    -fno-full-laziness
-
-  ghc-options:
-    -Wall
-    -Wnoncanonical-monad-instances
-    -Wincomplete-uni-patterns
-    -Wincomplete-record-updates
-    -Wredundant-constraints
-    -Widentities
-    -Wunused-packages
-    -Wmissing-deriving-strategies
-
-library bench-common
-  import:          lang
-  ghc-options:     -Wno-unused-packages
-  hs-source-dirs:  common
-  exposed-modules:
-    ZkFoldBenchmark.Common
-  build-depends:
-    base                         >=4.9 && <5,
-    bytestring                              ,
-    criterion                               ,
-    directory                               ,
-    filepath                                ,
-    flat                              ^>=0.6,
-    plutus-core                      ^>=1.24,
-    plutus-tx                        ^>=1.24,
-    tasty                                   ,
-    tasty-golden                            ,
-    temporary                               ,
-    text                                    ,
-    zkfold-base                             ,
-    zkfold-cardano                          ,
-
-library bench-lib
-  import:          lang
-  ghc-options:     -Wno-unused-packages
-  hs-source-dirs:  src
-  exposed-modules:
-    ZkFoldBenchmark.Verifier.RunVerifier
-    ZkFoldBenchmark.Verifier.Scripts
-  build-depends:
-    aeson                                   ,
-    base                         >=4.9 && <5,
-    bench-common                            ,
-    bytestring                              ,
-    hedgehog                                ,
-    plutus-core                      ^>=1.24,
-    plutus-tx                        ^>=1.24,
-    plutus-tx-plugin                 ^>=1.24,
-    zkfold-base                             ,
-    zkfold-cardano                          ,
-    containers                              ,
-
-executable bench-cpu-mem
-  import:           lang
-  default-language: Haskell2010
-  main-is:          Main.hs
-  hs-source-dirs:   exe
-  build-depends:
-    base                         >=4.9 && <5,
-    bench-lib                               ,
